Update poppy_all_apis test

+++ poppy_all_apis.py
@@ -16,6 +16,12 @@ class PoppyTasks(locust.TaskSet):
                                   "X-Auth-Token": CONFIG.token}
     service_ids = []
 
+    def __init__(self, *args, **kwargs):
+        super(PoppyTasks, self).__init__(*args, **kwargs)
+        # create a service so everything doesn't fail initially
+        print "Creating initial service"
+        self.post_service()
+
     @locust.task(CONFIG.create_service_weight)
     def post_service(self):
 
@@ -36,16 +42,22 @@ class PoppyTasks(locust.TaskSet):
 
         response = self.client.post('/'+self.tenant_id+'/services',
                                     data=json.dumps(post_data),
-                                    headers=self.headers)
+                                    headers=self.headers,
+                                    name="/{tenant}/services")
         print "Response for get is:", response.status_code
         print "Response content is:", response.content
-        print "Response extra content is: ", response.headers['location']
-        service_id = response.headers['location'].split('/')[-1]
-        print "Service id:", service_id
-        self.service_ids.append(service_id)
+        print "Response extra content is: ", response.headers.get('location')
+        if response.ok:
+            service_id = response.headers['location'].split('/')[-1]
+            print "Service id:", service_id
+            self.service_ids.append(service_id)
 
     @locust.task(CONFIG.update_service_weight)
     def update_service(self):
+        if not self.service_ids:
+            print "WARNING: service_ids list is empty"
+            return
+
         patch_data_update = [{
             "op": "add",
             "path": "/domains/-",
@@ -54,69 +66,84 @@ class PoppyTasks(locust.TaskSet):
                 "protocol": "http"
             }
         }]
-        service_id = self.service_ids.pop()
+        service_id = random.choice(self.service_ids)
         update_response = self.client.patch('/'+self.tenant_id+'/services/'
                                             +service_id,
                                             data=json.dumps(patch_data_update),
-                                            headers=self.headers)
-        self.service_ids.put(service_id)
+                                            headers=self.headers,
+                                            name="/{tenant}/services/{id}")
         print "Response for update is:", update_response.status_code
         print "Response content is:", update_response.content
-        print "Response xtra content is: ", update_response.headers['location']
 
     @locust.task(CONFIG.delete_service_weight)
     def delete_service(self):
+        if not self.service_ids:
+            print "WARNING: service_ids list is empty"
+            return
+
         service_id = self.service_ids.pop()
         delete_response = self.client.delete('/'+self.tenant_id+'/services/'
                                              +service_id,
-                                             headers=self.headers)
+                                             headers=self.headers,
+                                             name="/{tenant}/services/{id}")
         print "Response for delete is:", delete_response.status_code
         print "Response content is:", delete_response.content
-        print "Response xtra content is: ", delete_response.headers['location']
 
     @locust.task(CONFIG.delete_asset_weight)
     def delete_asset(self):
-        service_id = self.service_ids.pop()
+        if not self.service_ids:
+            print "WARNING: service_ids list is empty"
+            return
+
+        service_id = random.choice(self.service_ids)
         self.client.delete('/'+self.tenant_id+'/services/'+service_id
                            +'/assets',
                            headers=self.headers,
-                           params={'url': self._pick_asset()})
-        self.service_ids.put(service_id)
+                           params={'url': self._pick_asset()},
+                           name="/{tenant}/services/{id}/assets")
 
     @locust.task(CONFIG.delete_all_assets_weight)
     def delete_all_assets(self):
-        service_id = self.service_ids.pop()
+        if not self.service_ids:
+            print "WARNING: service_ids list is empty"
+            return
+        service_id = random.choice(self.service_ids)
         self.client.delete('/'+self.tenant_id+'/services/'+service_id
                            +'/assets',
                            headers=self.headers,
-                           params={'all': True})
-        self.service_ids.put(service_id)
+                           params={'all': True},
+                           name="/{tenant}/services/{id}/assets")
 
     @locust.task(CONFIG.list_services_weight)
     def list_services(self):
-        self.client.get('/'+self.tenant_id+'/services', headers=self.headers)
+        self.client.get('/'+self.tenant_id+'/services', headers=self.headers,
+                        name="/{tenant}/services")
 
     @locust.task(CONFIG.get_service_weight)
     def get_service(self):
-        service_id = self.service_ids.pop()
+        if not self.service_ids:
+            print "WARNING: service_ids list is empty"
+            return
+
+        service_id = random.choice(self.service_ids)
         self.client.get('/'+self.tenant_id+'/services/'+service_id,
-                        headers=self.headers)
-        self.service_ids.put(service_id)
+                        headers=self.headers, name="/{tenant}/services/{id}")
 
     @locust.task(CONFIG.list_flavors_weight)
     def list_flavors(self):
-        self.client.get('/'+self.tenant_id+'/flavors', headers=self.headers)
+        self.client.get('/'+self.tenant_id+'/flavors', headers=self.headers,
+                        name="/{tenant}/flavors")
 
     @locust.task(CONFIG.get_flavors_weight)
     def get_flavors(self):
         self.client.get('/'+self.tenant_id+'/flavors/'+self._pick_flavor(),
-                        headers=self.headers)
+                        headers=self.headers,
+                        name="/{tenant}/flavors/{flavor}")
 
     def _pick_flavor(self):
         return random.choice(CONFIG.flavors)
 
     def _pick_asset(self):
-    # TODO: idk my bff jill
         return '/index.html'
